chore: Remove ExecutorServiceFactoryProvider from ThreadPoolConfig
